refractor: Combine header tests into a single end-to-end test in widgetbook_test.dart
